fixup! chore: move upgrade-test js utils to lib/
