+++ packages/deployment/upgrade-test/upgrade-test-scripts/agoric-upgrade-12/mn2-start.test.js
@@ -37,12 +37,12 @@ import dbOpenAmbient from 'better-sqlite3';
 
 // TODO: factor out ambient authority from these
 // or at least allow caller to supply authority.
-import { mintIST } from '../lib/econHelpers.js/index.js';
-import { agoric, wellKnownIdentities } from '../lib/cliHelper.js/index.js';
+import { mintIST } from '../lib/econHelpers.js';
+import { agoric, wellKnownIdentities } from '../lib/cliHelper.js';
 import {
   provisionSmartWallet,
   voteLatestProposalAndWait,
-} from '../lib/commonUpgradeHelpers.js/index.js';
+} from '../lib/commonUpgradeHelpers.js';
 
 import { makeAgd } from '../lib/agd-lib.js';
 import { Far, makeMarshal, makeTranslationTable } from '../lib/unmarshal.js';
