+++ packages/deployment/upgrade-test/upgrade-test-scripts/lib/econHelpers.js
@@ -1,10 +1,6 @@
-import { agd, agops } from './lib/cliHelper.js/index.js';
-import {
-  ATOM_DENOM,
-  CHAINID,
-  VALIDATORADDR,
-} from './lib/constants.js/index.js';
-import { executeOffer } from './lib/commonUpgradeHelpers.js/index.js';
+import { agd, agops } from './lib/cliHelper.js';
+import { ATOM_DENOM, CHAINID, VALIDATORADDR } from './lib/constants.js';
+import { executeOffer } from './lib/commonUpgradeHelpers.js';
 
 export const openVault = (address, mint, collateral) => {
   return executeOffer(
